rename countryPaymentMethods to supportedPaymentMethods
